Создан LoginViewModel

+++ app/src/main/java/com/ulanapp/skywebprotestapp/domain/model/WeatherResponse.kt
@@ -22,7 +22,7 @@ data class WeatherResponse(
         @SerializedName("temp") var temp: Float? = null,
         @SerializedName("feels_like") var feelsLike: Float? = null,
         @SerializedName("temp_min") var tempMin: Float? = null,
-        @SerializedName("temp_max") var tempMax: Int? = null,
+        @SerializedName("temp_max") var tempMax: Float? = null,
         @SerializedName("pressure") var pressure: Int? = null,
         @SerializedName("humidity") var humidity: Int? = null
     )
